requirements.txt
pandas
numpy
tqdm
rapidfuzz
python-dotenv
openai
